pyproject.toml
[project]
name = "n31l"
version = "2.3.0"
description = "N31L is a utilitarian bot for the Call of Duty server."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "environs>=14.1.1",
    "loguru>=0.7.3",
    "loguru-discord>=1.4.0",
    "hikari-arc>=2.1.0",
    "httpx>=0.28.1",
    "asyncpraw>=7.8.1",
    "urlextract>=1.9.0",
    "uvloop>=0.21.0; sys_platform != 'win32'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.7",
    "asyncpraw-stubs>=0.0.3",
]

[tool.ruff.lint]
select = ["I"]
